+++ src/main/java/nextstep/courses/domain/attendee/Attendee.java
@@ -1,5 +1,6 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.attendee;
 
+import nextstep.courses.domain.session.Session;
 import nextstep.users.domain.NsUser;
 
 import java.util.Objects;
