+++ src/main/java/nextstep/courses/domain/attendee/AttendeeRepository.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.attendee;
 
 public interface AttendeeRepository {
