+++ src/main/java/nextstep/courses/domain/attendee/Attendees.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.attendee;
 
 import nextstep.courses.exception.AlreadyTakingSessionException;
