+++ src/main/java/nextstep/courses/domain/image/ImageInformation.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.image;
 
 import nextstep.courses.exception.ExceedLimitedVolumeException;
