+++ src/main/java/nextstep/courses/domain/session/Session.java
@@ -1,5 +1,8 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
+import nextstep.courses.domain.image.Image;
+import nextstep.courses.domain.attendee.Attendee;
+import nextstep.courses.domain.attendee.Attendees;
 import nextstep.payments.domain.Payment;
 import nextstep.users.domain.NsUser;
