+++ src/main/java/nextstep/courses/domain/session/TotalAttendee.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import nextstep.courses.exception.ExceedAttendeesException;
 import nextstep.courses.exception.NegativeNumberException;
