+++ src/test/java/nextstep/courses/domain/AttendeesTest.java
@@ -1,5 +1,8 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.attendee.Attendee;
+import nextstep.courses.domain.attendee.Attendees;
+import nextstep.courses.domain.session.*;
 import nextstep.courses.exception.AlreadyTakingSessionException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -17,7 +20,7 @@ class AttendeesTest {
     void throw_exception_if_user_already_enrolled_session() {
         Period period = new Period(LocalDateTime.now(), LocalDateTime.now().plusHours(1));
         SessionInformation information = new SessionInformation(SessionStatus.RECRUITING, SessionType.FREE, period);
-        Session session = new FreeSession(1L, information,null);
+        Session session = new FreeSession(1L, information, null);
         Attendee attendee = new Attendee(JAVAJIGI, session);
         Attendees attendees = new Attendees(List.of(attendee));
