+++ src/test/java/nextstep/courses/domain/ImageInformationTest.java
@@ -1,5 +1,8 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.image.ImageFormat;
+import nextstep.courses.domain.image.ImageInformation;
+import nextstep.courses.domain.image.ImageSize;
 import nextstep.courses.exception.ExceedLimitedVolumeException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
