+++ src/test/java/nextstep/courses/domain/ImageSizeTest.java
@@ -1,5 +1,6 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.image.ImageSize;
 import nextstep.courses.exception.NotValidSizeException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
